requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
